fix: correct emailverification package path from mailsend to mailersend
