+++ src/main/java/com/mailersend/sdk/emailverification/AsyncEmailVerificationResponse.java
@@ -5,7 +5,7 @@
  * @author MailerSend <support@mailersend.com>
  * https://mailersend.com
  **************************************************/
-package com.mailsend.sdk.emailverification;
+package com.mailersend.sdk.emailverification;
 
 import com.google.gson.annotations.SerializedName;
 import com.mailersend.sdk.MailerSendResponse;
