+++ src/main/java/com/mailersend/sdk/emailverification/EmailVerificationList.java
@@ -5,7 +5,7 @@
  * @author MailerSend <support@mailersend.com>
  * https://mailersend.com
  **************************************************/
-package com.mailsend.sdk.emailverification;
+package com.mailersend.sdk.emailverification;
 
 import java.time.Instant;
 import java.time.format.DateTimeFormatter;
